Split files into different .cs by SOLID principles
